% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmarketing.R
\docType{package}
\name{bmarketing}
\alias{bmarketing}
\alias{bmarketing-package}
\title{bmarketing: A package for analyzing datasets of banking customers.}
\description{
The bmarketing package provides three important functions:
clean, transform and calcPerformance.
}
\section{bmarketing functions}{
 \cr
- clean: A function to clean data (clean NA values, basic checks)\cr
- transform: A function to log transform values\cr
- calcPerformance: A function to ...
}

