% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNA.R
\name{checkNA}
\alias{checkNA}
\title{Checking for NA values considering all the variables present in the dataset}
\usage{
checkNA(ds)
}
\arguments{
\item{ds}{dataframe object}
}
\value{
Return an error message indicating how many columns have NAs present, in case there is any.
}
\description{
Checking for NA values considering all the variables present in the dataset
}
