% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Checking the quality of data.}
\usage{
cleanData(ds, targetVar)
}
\arguments{
\item{ds}{dataframe object}

\item{targetVar}{name of the target variable of the dataframe \code ds}
}
\value{
Return the cleaned dataframe and prints warning messages in case there are any.
}
\description{
NA values checked.
}
