% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAR.R
\name{clear1}
\alias{clear1}
\title{Function to find missing data}
\usage{
clear1(data)
}
\arguments{
\item{data}{Basic object like \code{numeric,int,boolean,char}}

\item{target}{Basic object like \code{numeric,int,boolean,char}}
}
\description{
Function to find missing data
}
\examples{
clear1(bmarketing$y)
clear2(bmarketing$duration)
clear3(bmarketing$poutcome)

}
\keyword{,}
\keyword{length}
\keyword{na.omit}
\keyword{nrow,}
