% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAR.R
\name{clear2}
\alias{clear2}
\title{Function to find missing data inside a column}
\usage{
clear2(data, target)
}
\arguments{
\item{data}{Basic object like \code{numeric,int,boolean,char}}

\item{target}{Basic object like \code{numeric,int,boolean,char}}
}
\description{
Function to find missing data inside a column
}
\examples{
clear2(bmarketing$duration)

}
\keyword{,}
\keyword{length}
\keyword{na.omit}
\keyword{nrow,}
