% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAR.R
\name{clear3}
\alias{clear3}
\title{Remove any columns (and report as warning) which contain more than 50 percent NA’s}
\usage{
clear3(data, target)
}
\arguments{
\item{data}{Basic object like \code{numeric,int,boolean,char}}

\item{target}{Basic object like \code{numeric,int,boolean,char}}
}
\description{
Remove any columns (and report as warning) which contain more than 50 percent NA’s
}
\examples{
clear3(bmarketing$duration)

}
\keyword{length}
\keyword{na.omit,}
\keyword{nrow,}
