% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtree}
\alias{dtree}
\title{dtree}
\usage{
dtree(udata, target)
}
\arguments{
\item{udata}{input data for the model.}

\item{target}{the name of the target variable from the input data (string) for example: target_name="Y"}
}
\value{
A decision tree model.
}
\description{
fits the decision tree model based on given parameters
}
