% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtreeperf}
\alias{dtreeperf}
\title{dtreeperf}
\usage{
dtreeperf(target, predictions)
}
\arguments{
\item{target}{actual target in our data}

\item{predictions}{predicted target}
}
\value{
accuracy
}
\description{
checks model accuracy
}
