% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtreeplot}
\alias{dtreeplot}
\title{dtreeplot}
\usage{
dtreeplot(dt_model)
}
\arguments{
\item{dt_model}{model that we want to plot}
}
\value{
a plot. :)
}
\description{
returns the plot of the chosen decision tree model
}
\examples{
dtreeplot(mymodel)

}
