% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtreepredict}
\alias{dtreepredict}
\title{dtreepredict}
\usage{
dtreepredict(dt_model, predictdata)
}
\arguments{
\item{dt_model}{model that we want to use to generate our predictions}

\item{predictdata}{data that we want to score}
}
\value{
predictions
}
\description{
returns the prediction
}
