% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtreesummary}
\alias{dtreesummary}
\title{dtreesummary}
\usage{
dtreesummary(dt_model)
}
\arguments{
\item{dt_model}{model that we want to summarize}
}
\value{
a summary. :)
}
\description{
returns the summary of the chosen model
}
\examples{
dtreesummary(mymodel)

}
