% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Reports model's classification accuracy measures}
\usage{
fitModel(data, y, modelType, explVars = NULL)
}
\arguments{
\item{\code{data}}{data.frame used for model fitting}

\item{\code{y}}{name of the target variable (quoted character)}

\item{\code{modelType}}{string, name of the requested model type: either 'Logistic' or 'DecisionTree'}

\item{\code{explVars}}{either NULL or character vector containing list of explanatory variables}
}
\value{
Return the objectum of the model
}
\description{
Reports model's classification accuracy measures
}
\examples{
df <- data.frame(y = c(0,1,1,0), a = c('a', 'b', 'c', 'a'), b = c(12,121,11,12))
varList <- c('a','b')
results <- fitModel(data = df, y = 'y', modelType = 'Logistic', explVars = varList)
}
