% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{histplot}
\alias{histplot}
\title{Function to histogram}
\usage{
histplot(data, column)
}
\arguments{
\item{data}{Basic object like \code{numeric, char, factor,boolean, NULL}}

\item{column}{Basic object like \code{numeric}}
}
\description{
Function to histogram
}
\examples{
histplot(data, column)


}
\keyword{plot}
