% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_from_csv}
\alias{import_data_from_csv}
\title{Imports Data from a csv file}
\usage{
import_data_from_csv(filename, ...)
}
\arguments{
\item{filename}{character, filename}

\item{...}{args, forward to read_csv function}
}
\value{
data.frame
}
\description{
Imports Data from a csv file
}
