% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{intonum}
\alias{intonum}
\title{Transform factors into numeric variables (and vice versa) as necessary}
\usage{
intonum(data, column)
}
\arguments{
\item{data}{Basic object like \code{numeric, boolean, character, NULL, factor}}

\item{column}{Basic object like \code{factor}}
}
\description{
Transform factors into numeric variables (and vice versa) as necessary
}
\examples{
bmarketing$poutcome<-intonum(bmarketing$poutcome)
bmarketing$euribor3m<-intonum(bmarketing, bmarketing$euribor3m)

}
\keyword{number2factor}
