% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{logaritmic}
\alias{logaritmic}
\title{Transform numeric variables using the log}
\usage{
logaritmic(x)
}
\arguments{
\item{x}{Basic object like \code{numeric}}

\item{data}{Basic object like \code{numeric, boolean, character, NULL, factor}}

\item{column}{Basic object like \code{factor}}
}
\description{
Transform numeric variables using the log
}
\examples{
bmarketing$age<-logaritmic(bmarketing$age)
bmarketing$duration<-logaritmic(bmarketing$duration)

}
\keyword{log}
