% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(udata, target)
}
\arguments{
\item{udata}{Input data for the model.}

\item{target}{target variable from the input data.}
}
\value{
logit model.
}
\description{
fits the logit model based on
}
