% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Function to model}
\usage{
model(dataset, target)
}
\arguments{
\item{dataset}{Basic object like \code{numeric, char, factor,boolean, NULL}}

\item{target}{Basic object like \code{boolean}}
}
\description{
Function to model
}
\examples{
model(bmarketing, bmarketing$y)


}
\keyword{decision}
\keyword{model,}
\keyword{tree,}
