% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Returns the predicted classes from the decision tree}
\usage{
model_data(df, target_var)
}
\arguments{
\item{df}{data.frame on which we want to make a model}

\item{target_var}{character, Dependent variable}
}
\value{
return decision tree model
}
\description{
Returns the predicted classes from the decision tree
}
\examples{
\dontrun{
model_data(df, target_var="Y")
}
}
