% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{Plot the tree and respective nodes}
\usage{
model_plot(dt_model)
}
\arguments{
\item{dt_model}{randomForest, received by \code{link{model_data}}}
}
\description{
Plot the tree and respective nodes
}
\examples{
\dontrun{
model_plot(dt_model)
}
}
