% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{model_pred}
\alias{model_pred}
\title{Function to model predict}
\usage{
model_pred(dataset, dt_model)
}
\arguments{
\item{dataset}{Basic object like \code{numeric, char, factor,boolean, NULL}}

\item{dt_model}{Basic object like \code{object}}
}
\description{
Function to model predict
}
\examples{
model_pred(bmarketing, model)


}
\keyword{decision}
\keyword{model,}
\keyword{prediction}
\keyword{tree,}
