% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{model_predict}
\alias{model_predict}
\title{Title}
\usage{
model_predict(dt_model, data, .)
}
\arguments{
\item{dt_model, }{randomForest, received by \code{link{model_data}}}

\item{data}{df data.frame , received by \code{\link{clean_data}}}

\item{.}{}
}
\description{
The main goal of linear regression is 
to predict an outcome value on the basis of one or multiple predictor variables.
}
\examples{


}
