% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictByModel.R
\name{predictByModel}
\alias{predictByModel}
\title{Predicts output from DecisitonTree or Logistic model}
\usage{
predictByModel(data, model2Predict, modelType)
}
\arguments{
\item{data}{dataset to be used for prediction}

\item{model2Predict}{model which the predictions are based on}

\item{modelType}{type of the model}
}
\value{
Gives back predictions
}
\description{
Predicts output from DecisitonTree or Logistic model
}
\examples{
y_example      = c(0,1,1,0)
y_pred_example = c(1,1,1,0)
}
