% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Predictions.}
\usage{
predictions(dt_model, bmarketing)
}
\arguments{
\item{dt_model}{model}

\item{bmarketing}{data}
}
\description{
\code{predictions} returns predictions.
}
\examples{
predictions(model, data)
}
