\name{Team1-package}
\alias{Team1-package}
\alias{Team1}
\docType{package}
\title{
\packageTitle{Team1}
}
\description{
\packageDescription{Team1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Team1}
\packageIndices{Team1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Team1}

Maintainer: \packageMaintainer{Team1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
