% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmarketing.R
\docType{data}
\name{bmarketing}
\alias{bmarketing}
\title{Banking Marketing Data}
\format{A data frame with 4119 rows and 21 variables:
\describe{
  \item{age}{age, integer}
  \item{job}{factor}
  \item{marital}{factor}
  \item{education}{factor}
  \item{default}{factor}
  \item{housing}{factor}
  \item{load}{factor}
  \item{contact}{factor}
  \item{month}{factor}
  \item{day_of_week}{factor}
  \item{duration}{integer}
  \item{campaign}{integer}
  \item{pdays}{integer}
  \item{previous}{integer}
  \item{poutcome}{factor}
  \item{emp.var.rate}{numeric}
  \item{cons.price.idx}{numeric}
  \item{cons.conf.idx}{numeric}
  \item{euribor3m}{numeric}
  \item{nr.employed}{numeric}
  \item{y}{target variable}
}}
\source{
marketing
}
\usage{
bmarketing
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
