% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerformance.R
\name{calcPerformance}
\alias{calcPerformance}
\title{Reports model's classification accuracy measures}
\usage{
calcPerformance(y, y_pred)
}
\arguments{
\item{\code{y}}{Target variable (class or numeric)}

\item{\code{y}}{Target variable (class or numeric)}
}
\value{
Gives back a classification report containing the Confusion Matrix, Sensitivity, Specificity, Precision and Accuracy
}
\description{
Reports model's classification accuracy measures
}
\examples{
calcDTPerformance(y = bmarketing$y, y_pred = predictions)
}
