% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerformance.R
\name{calcPerformance}
\alias{calcPerformance}
\title{Reports model's classification accuracy measures}
\usage{
calcPerformance(y, y_pred)
}
\arguments{
\item{\code{y}}{Target variable (class or numeric)}

\item{\code{y_pred}}{Target variable (class or numeric)}
}
\value{
Gives back a classification report containing the Confusion Matrix, Sensitivity, Specificity, Precision and Accuracy, and a list containing the results
}
\description{
Reports model's classification accuracy measures
}
\examples{
y_example      = c(0,1,1,0)
y_pred_example = c(1,1,1,0)
results <- calcPerformance(y = y_example, y_pred = y_pred_example)
}
