% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Cleanin Data}
\usage{
clean(data, target_name)
}
\arguments{
\item{data}{A data.frame containing a target variable named y}

\item{target_name}{A character containing the name of the target variable}
}
\value{
A cleaned data.frame. 

This means:

  - return an error if the target variable contains any missing values (NA’s).
  - Give clear warnings for all other variables which contain NA’s.
  - Remove any columns (and report as warning) which contain more than 50% NA’s.
}
\description{
\code{clean} returns a clean data set.
}
\examples{
data(bmarketing)
clean(bmarketing,"y")
}
