% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean function.}
\usage{
clean(x, t)
}
\arguments{
\item{x}{A dataframe}

\item{t}{The name of the target variable column of dataframe x}
}
\description{
Cleans a dataset: \cr
- return an error if the target variable contains any missing values (NA’s). \cr
- Give clear warnings for all other variables which contain NA’s. \cr
- Remove any columns (and report as warning) which contain more than 50% NA’s
}
