% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtree}
\alias{dtree}
\title{dtree}
\usage{
dtree(target, udata)
}
\arguments{
\item{target}{target variable from the input data.}

\item{udata}{input data for the model.}
}
\value{
A decision tree model.
}
\description{
fits the decision tree model based on given parameters
}
