% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{logaritmic}
\alias{logaritmic}
\title{Function to make neccessary transformation of data}
\usage{
logaritmic(x)
}
\arguments{
\item{x}{Basic object like \code{numeric}}

\item{data}{Basic object like \code{numeric, boolean, character, NULL, factor}}

\item{column}{Basic object like \code{factor}}
}
\description{
Function to make neccessary transformation of data
}
\examples{
bmarketing$age<-logaritmic(bmarketing$age)
bmarketing$duration<-logaritmic(bmarketing$duration)
bmarketing$poutcome<-intonum(bmarketing$poutcome)
bmarketing$euribor3m<-intonum(bmarketing, bmarketing$euribor3m)

}
\keyword{,}
\keyword{log}
\keyword{number2factor}
