% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{logaritmic}
\alias{logaritmic}
\title{Function to make neccessary transformation of data}
\usage{
logaritmic(x)
}
\arguments{
\item{x}{Basic object like \code{numeric}}

\item{f}{Basic object like \code{factor}}
}
\description{
Function to make neccessary transformation of data
}
\examples{
logaritmic(bmarketing$age)
logaritmic(bmarketing$duration)
intonum(bmarketing$poutcome)
intonum(bmarketing$euribor3m)

}
\keyword{,}
\keyword{log}
\keyword{number2factor}
