% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Runing decision tree model and ploting it results}
\usage{
model(input_data, target_name)
}
\arguments{
\item{data}{A data.frame containing a target variable named which name is defined as input argument}
}
\value{
A decision tree model 

This means:

  Model: Create a decision tree model to predict if customer signing a term deposit.
Model-Plot: We shall implement a function to present a nice representation of the model, e.g. for a decision tree we should plot the tree and respective nodes.
}
\description{
\code{clean} returns a decision tree model.
}
\examples{
data("bmarketing")
dt_model <- model(input_data = bmarketing,target_name="y")
}
