% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_accuracy.R
\name{model_accuracy}
\alias{model_accuracy}
\title{model_accuracy}
\usage{
model_accuracy(real, pred)
}
\arguments{
\item{real}{a vector of realized observations}

\item{pred}{a vector of corresponding predictions}
}
\value{
\code{accuracy} ... a number

\code{confusion_matrix} ... a matrix
}
\description{
The function calculates accuracy (as a share of correctly classified observations and total observations) and confusion matrix of a classification model. For further reference to confusion matrix please see \url{https://en.wikipedia.org/wiki/Confusion_matrix}.
}
\examples{
example_real=c(1,2,3,1,2,3,1,2,3) # a vector of realized observations
example_pred=c(1,2,3,1,2,3,2,2,2) # a vector of predictions
model_accuracy(example_real,example_pred)
}
