% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Calculate the accuracy of the model}
\usage{
model_performance(data, prediction)
}
\arguments{
\item{data}{character vector, real results}

\item{prediction}{character vector, predicted results}
}
\value{
list, accuracy measures out of confusion matrix:
}
\description{
Calculate the accuracy of the model
}
