% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Function to model performance}
\usage{
model_performance(target, predictions)
}
\arguments{
\item{target}{Basic object like \code{boolean}}

\item{predictions}{Basic object like \code{boolean}}
}
\description{
Function to model performance
}
\examples{
model_performance(target, predictions)


}
\keyword{accuracy}
\keyword{decision}
\keyword{model,}
\keyword{tree,}
