% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Calculate the accuracy of the model}
\usage{
model_performance(target, prediction)
}
\arguments{
\item{target}{vector, target values}

\item{prediction}{vector, predicted values}
}
\value{
list, measures out of confusion matrix: accuracy, sensitivity, specificity. \url{https://en.wikipedia.org/wiki/Confusion_matrix}
}
\description{
Calculate the accuracy of the model
}
\examples{
target <- sample(c(0, 1), 100, replace = TRUE)
prediction <- sample(c(0, 1), 100, replace = TRUE)
model_performance(target, prediction)

}
