% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Predictions.}
\usage{
predictions(dt_model, bmarketing)
}
\description{
\code{predictions} returns predictions.
}
\examples{
data("bmarketing")
dt_model <- model(input_data = bmarketing,target_name="y")
predictions(dt_model, bmarketing)
}
