% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Log transformation and factor transformation into numeric variables (and vice versa) as necessary.}
\usage{
transform(dataset)
}
\description{
This function applies logarithm on the sent values and transforms numerical data to categories and vice versa
In case of null values, these are replaced with the mean of the interval for numerical data.
}
