% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{transform}
\alias{transform}
\title{Log-Tranforms input data for easy model development\cr}
\usage{
transform(input)
}
\arguments{
\item{input}{dataframe}
}
\description{
If data is numeric and negative \cr
AND if data appears as non-normal with the shapiro test\cr
the function performs a log transformation\cr
}
