% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Tranform data}
\usage{
transform_data(df, cols)
}
\arguments{
\item{df}{data.frame, recveived by \code{\link{clean_data}}}

\item{cols}{character vector, columns to be log-transformed}
}
\value{
data.frame
}
\description{
1. Changes factor columns to numeric, if possible
2. Apply log to user defined columns
}
